dataset_scripts/requirements.txt
tqdm>=4.64.0
Pillow>=9.3.0
requests>=2.28.0
